+++ tests/test_netcdf_convert.py
@@ -17,7 +17,6 @@ import xarray as xr
 from net2cog.netcdf_convert import (
     Net2CogError,
     get_all_data_variables,
-    has_spatial_dimensions,
     netcdf_converter,
 )
 
@@ -147,73 +146,7 @@ def test_excluded_variables_not_converted(temp_dir, logger, smap_file):
     assert basename(results[0]) == 'gland.tif', 'Incorrect output file name'
 
 
-@pytest.mark.parametrize(
-    'dimensions',
-    [['lat', 'lon'], ['latitude', 'longitude'], ['x', 'y'], ['x-dim', 'y-dim']],
-)
-def test_has_spatial_dimensions_present(dimensions):
-    """Verify returns True for variable with spatial dimensions."""
-    test_datatree = xr.DataTree(
-        dataset=xr.Dataset(
-            data_vars={'science': ([dimensions[0], dimensions[1]], np.ones((2, 4)))},
-            coords={
-                dimensions[0]: (dimensions[0], np.array([1, 2])),
-                dimensions[1]: (dimensions[1], np.array([3, 4, 5, 6])),
-            },
-        ),
-    )
-    assert has_spatial_dimensions(test_datatree['science'])
-
-
-@pytest.mark.parametrize(
-    'dimensions',
-    [['lat', 'lon'], ['latitude', 'longitude'], ['x', 'y'], ['x-dim', 'y-dim']],
-)
-def test_has_spatial_dimensions_and_others_present(dimensions):
-    """Verify returns True, when spatial dimensions and others are present."""
-    test_datatree = xr.DataTree(
-        dataset=xr.Dataset(
-            data_vars={
-                'science': (['time', dimensions[0], dimensions[1]], np.ones((1, 2, 4)))
-            },
-            coords={
-                'time': ('time', np.array([0])),
-                dimensions[0]: (dimensions[0], np.array([1, 2])),
-                dimensions[1]: (dimensions[1], np.array([3, 4, 5, 6])),
-            },
-        ),
-    )
-    assert has_spatial_dimensions(test_datatree['science'])
-
-
-@pytest.mark.parametrize('dimension', ['lat', 'latitude', 'x', 'x-dim'])
-def test_has_spatial_dimensions_incomplete(dimension):
-    """Verify returns False when only one spatial dimension present."""
-    test_datatree = xr.DataTree(
-        dataset=xr.Dataset(
-            data_vars={'science': ([dimension], np.ones((4)))},
-            coords={
-                dimension: (dimension, np.array([1, 2, 3, 4])),
-            },
-        ),
-    )
-    assert not has_spatial_dimensions(test_datatree['science'])
-
-
-def test_has_spatial_dimensions_absent():
-    """Verify returns False for variable without spatial dimensions."""
-    test_datatree = xr.DataTree(
-        dataset=xr.Dataset(
-            data_vars={'science': (['time'], np.ones(4))},
-            coords={
-                'time': ('time', np.array([1, 2, 3, 4])),
-            },
-        ),
-    )
-    assert not has_spatial_dimensions(test_datatree['science'])
-
-
-def test_get_all_data_variables_flat_input():
+def test_get_all_data_variables_flat_input(logger):
     """Verify returns all data variables from a file with a single root group."""
     test_datatree = xr.DataTree(
         dataset=xr.Dataset(
@@ -230,12 +163,12 @@ def test_get_all_data_variables_flat_input():
             },
         ),
     )
-    assert set(get_all_data_variables(test_datatree)) == set(
+    assert set(get_all_data_variables(test_datatree, logger)) == set(
         ['/science_one', '/science_two', '/science_three']
     )
 
 
-def test_get_all_data_variables_hierarchical_input():
+def test_get_all_data_variables_hierarchical_input(logger):
     """Verify returns all data variables from a file with nested groups.
 
     Tree structure in test:
@@ -275,7 +208,7 @@ def test_get_all_data_variables_hierarchical_input():
         ),
     )
 
-    assert set(get_all_data_variables(test_datatree)) == set(
+    assert set(get_all_data_variables(test_datatree, logger)) == set(
         ['/science_one', '/group_one/science_two', '/group_one/group_two/science_three']
     )
 
@@ -358,3 +291,58 @@ def test_spl3smp_nested_variable_3d_annotated(
         basename(results[0]) == "Soil_Moisture_Retrieval_Data_AM_landcover_class.tif"
     ), "Incorrect output file name"
     assert cog_validate(pathlib.Path(results[0]))[0]
+
+
+def test_spl3smp_all_variable_3d_annotated(
+    temp_dir, logger, spl3smp_nested_3d_annotated_file
+):
+    """Verify a SPL3SMP all variable generate COG for supported dtype
+    [ubyte|uint8|uint16|int16|uint32|int32|float32|float64] only.
+
+    The string variables tb_time_utc_am and tb_time_utc_pm will not
+    generate COG files."
+
+    """
+    test_file = pathlib.Path(temp_dir, spl3smp_nested_3d_annotated_file)
+
+    results = netcdf_converter(
+        test_file,
+        pathlib.Path(temp_dir),
+        [],
+        logger,
+    )
+
+    # Check results are as expected:
+    assert len(results) == 8, "Incorrect number of output file names."
+
+    for entry in results:
+        if pathlib.Path(entry).is_file():
+            assert (
+                basename(entry[0])
+                != "Soil_Moisture_Retrieval_Data_AM_tb_time_utc_am.tif"
+            )
+            assert (
+                basename(entry[0])
+                != "Soil_Moisture_Retrieval_Data_AM_tb_time_utc_pm.tif"
+            )
+            assert cog_info(pathlib.Path(entry)).GEO.CRS == "EPSG:6933"
+            assert cog_validate(pathlib.Path(entry))
+
+
+def test_spl3smp_dtype_string_handle_exception(
+    temp_dir, logger, spl3smp_nested_3d_annotated_file
+):
+    """Verify a SPL3SMP variable with dtype=string (S1) throws exception"""
+    test_file = pathlib.Path(temp_dir, spl3smp_nested_3d_annotated_file)
+    expected_exception = (
+        "Variable /Soil_Moisture_Retrieval_Data_AM/tb_time_utc"
+        " cannot be converted to tif: invalid dtype: dtype\\('S1'\\)"
+    )
+
+    with pytest.raises(Net2CogError, match=expected_exception):
+        netcdf_converter(
+            test_file,
+            pathlib.Path(temp_dir),
+            ["/Soil_Moisture_Retrieval_Data_AM/tb_time_utc"],
+            logger,
+        )